feat: skip JS build when pre-built dist assets exist (get-app + update)
Moves asset-handling logic into PythonEnvManager.build_assets_for_app():

- Checks for esbuild content-hash files in apps/{name}/{name}/public/dist/js/
- If found: creates sites/assets/{app}/ symlink and generates assets.json
  + assets-rtl.json by scanning dist/ — no Node.js or yarn required
- If not found: falls back to yarn install + frappe build --force

get-app delegates to this helper (replacing the inline subprocess calls).
bench update gains a new _rebuild_assets() step that calls the same
helper for every app after reinstalling Python packages.
